+++ common-model-repository/change-log.md
@@ -1,3 +1,9 @@
+# 0.1.11 2021-08-17
+    1. 加入Operator接口，QueryOperator,SortOperator,LogicOperator都实现该接口
+    
+# 0.1.10 2021-07-26
+    1. QueryOperator加入LK
+    
 # 0.1.9 2020-12-12
     1.加入PropertyMapping.getPropertyFullName方法，返回带全路径的属性名(user.id)